Move macro definition of "make_bitfield_serde" to struct_accessors.
Fixes <https://github.com/oxidecomputer/amd-efs/issues/124>.
